+++ splash.js
@@ -9,7 +9,7 @@
 * First License: GPLv2
 * Github: https://github.com/cameramanben/LUTCalc
 */
-lutInputs.addInput('version','v2.5');
+lutInputs.addInput('version','v2.53');
 lutInputs.addInput('date','August 2016');
 var splash = document.getElementById('javascriptwarning');
 splash.removeChild(splash.firstChild);