+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 LUTCalc Changelog
 =================
 
+v2.53
+-----
+* New Feature - REDWideGamutRGB and Log3G10.
+
+v2.52
+-----
+* Bugfix - Fixed Canon C-Log3 bug.
+
+v2.51
+-----
+* New Feature - Canon C-Log3.
+
 v2.5
 ----
 * New Feature - Black Gamma.